Open: -g as a primitive "fuzzy search" feature.
XDel is rarely used, and having to search & click for a file
in +Files is cumbersome, so the experiment of
f49492e6bd4aaee1b53f530f0e105fb3c5d55f5b brings little.

Open OTOH is used more often, and provides a convenient interface
for opening a ("common") file matching a pattern. We could be
smarter:

  - regularly dump all opened files to $HOME/acme.files (the
  "new" +Files buffer that isn't meant to be interacted with
  manually) instead of relying on Open to update it;

  - interact with existing fuzzy finder;

  - select which file to open when more than one pattern match
  by e.g. selecting the one most often opened, or having an
  option to display them all.